# ai/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CandleItem": {
        "properties": {
          "close": {
            "title": "Close",
            "type": "integer"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "high": {
            "title": "High",
            "type": "integer"
          },
          "low": {
            "title": "Low",
            "type": "integer"
          },
          "open": {
            "title": "Open",
            "type": "integer"
          },
          "volume": {
            "title": "Volume",
            "type": "integer"
          }
        },
        "required": [
          "date",
          "open",
          "high",
          "low",
          "close",
          "volume"
        ],
        "title": "CandleItem",
        "type": "object"
      },
      "CandlePeriod": {
        "enum": [
          "1M",
          "3M",
          "1Y"
        ],
        "title": "CandlePeriod",
        "type": "string"
      },
      "CandlesResponse": {
        "properties": {
          "candles": {
            "items": {
              "$ref": "#/components/schemas/CandleItem"
            },
            "title": "Candles",
            "type": "array"
          },
          "interval": {
            "default": "DAY",
            "title": "Interval",
            "type": "string"
          },
          "period": {
            "$ref": "#/components/schemas/CandlePeriod"
          },
          "stockCode": {
            "title": "Stockcode",
            "type": "string"
          }
        },
        "required": [
          "stockCode",
          "period",
          "candles"
        ],
        "title": "CandlesResponse",
        "type": "object"
      },
...
      "DailyCloseItem": {
        "properties": {
          "close": {
            "title": "Close",
            "type": "integer"
          },
          "stockCode": {
            "title": "Stockcode",
            "type": "string"
          }
        },
        "required": [
          "stockCode",
          "close"
        ],
        "title": "DailyCloseItem",
        "type": "object"
      },
      "DailyCloseResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DailyCloseItem"
            },
            "title": "Items",
            "type": "array"
          },
          "tradeDate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tradedate"
          }
        },
        "required": [
          "tradeDate",
          "items"
        ],
        "title": "DailyCloseResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "internalToken": {
        "description": "서비스 간 공유 토큰. 호출자가 백엔드임을 확인한다. 이 서버는 외부에 노출되지 않는 전제이며, 토큰은 그 문을 잠그는 두 번째 자물쇠다.",
        "in": "header",
        "name": "X-Internal-Token",
        "type": "apiKey"
      },
      "trustedUserHeader": {
        "description": "백엔드가 JWT를 검증한 뒤 넣어 주는 사용자 식별자. 경로나 본문으로는 받지 않는다. AI 서버는 이 값을 검증 없이 신뢰하므로 내부 네트워크에서만 접근 가능해야 한다.",
        "in": "header",
        "name": "X-User-Id",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "국내 주식 전용. 수치는 엔진에서만 나오고 LLM은 설명만 생성한다.",
    "title": "AI 투자 비서 — AI 파트",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/prices/daily-close": {
      "get": {
        "description": "요청일 또는 가장 최근 거래일의 전 종목 종가를 돌려준다.",
        "operationId": "get_daily_close_internal_prices_daily_close_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyCloseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "internalToken": []
          }
        ],
        "summary": "Get Daily Close",
        "tags": [
          "internal-prices"
        ]
      }
    },
    "/internal/prices/{stock_code}/candles": {
      "get": {
        "description": "KST 현재일을 기준으로 수정주가 일봉을 오래된 순서부터 돌려준다.",
        "operationId": "get_candles_internal_prices__stock_code__candles_get",
        "parameters": [
          {
            "in": "path",
            "name": "stock_code",
            "required": true,
            "schema": {
              "pattern": "^\\d{6}$",
              "title": "Stock Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "period",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CandlePeriod"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandlesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "internalToken": []
          }
        ],
        "summary": "Get Candles",
        "tags": [
          "internal-prices"
        ]
      }
    }
  },
  "security": [
    {
      "internalToken": [],
      "trustedUserHeader": []
    }
  ],
  "servers": [
    {
      "description": "로컬 개발 서버",
      "url": "http://localhost:8000"
    }
  ]
}